fix: get rid of warnings

+++ src/monitor/metrics.c
@@ -18,7 +18,7 @@ static user_info active_users[MAX_USERS];
 static int user_count = 0;
 
 static int timeout_seconds = 30;
-static int max_connections = 1;
+static uint64_t max_connections = 1;
 
 void metrics_init(struct selector_init *conf) {
     config = conf;
